+++ 0x0C-python-almost_a_circle/tests/test_models/test_square.py
@@ -21,18 +21,12 @@ class SquareTest(unittest.TestCase):
         Base._Base__nb_objects = 0
 
     def tearDown(self):
-        try:
-            os.remove("Rectangle.json")
-        except:
-            pass
-        try:
+        if os.path.exists("Base.json"):
             os.remove("Base.json")
-        except:
-            pass
-        try:
+        if os.path.exists("Rectangle.json"):
+            os.remove("Rectangle.json")
+        if os.path.exists("Square.json"):
             os.remove("Square.json")
-        except:
-            pass
 
     def test_00_one_arg(self):
         """Test for one argument passed in."""
